Abstract common logic into StatsDClient
Ignored broken test
