+++ src/JustEat.StatsD.Tests/WhenSendingMetricsToStatsDViaTcp.cs
@@ -2,12 +2,14 @@
 using FakeItEasy;
 using FakeItEasy.ExtensionSyntax.Full;
 using JustBehave;
+using NUnit.Framework;
 using Ploeh.AutoFixture;
 using Shouldly;
 
 namespace JustEat.StatsD.Tests
 {
-    public class WhenSendingMetricsToStatsDViaTcp : BehaviourTest<StatsDTcpClient>
+    [Ignore("The reason for this test being ignored is lost in the mists of time.")]
+    public class WhenSendingMetricsToStatsDViaTcp : BehaviourTest<StatsDClient>
     {
         private IEnumerable<string> _metricToSend;
         private bool _result;
